requirements.txt
numpy
scipy
networkx
matplotlib
tqdm
pyyaml
